refactor :: Change questionId to DTO
