+++ src/main/kotlin/hs/kr/equus/feed/infrastructure/kafka/consumer/reply/ReplyCreatedEventConsumer.kt
@@ -2,7 +2,7 @@ package hs.kr.equus.feed.infrastructure.kafka.consumer.reply
 
 import hs.kr.equus.feed.domain.question.service.MarkAsRepliedService
 import hs.kr.equus.feed.infrastructure.kafka.configuration.KafkaTopics
-import hs.kr.equus.user.infrastructure.kafka.dto.CreateReplyEventRequest
+import hs.kr.equus.feed.infrastructure.kafka.dto.CreateReplyEventRequest
 import org.springframework.kafka.annotation.KafkaListener
 import org.springframework.stereotype.Service
 
@@ -10,7 +10,11 @@ import org.springframework.stereotype.Service
 class ReplyCreatedEventConsumer(
     private val markAsRepliedService: MarkAsRepliedService
 ) {
-    @KafkaListener(topics = [KafkaTopics.CREATE_REPLY], groupId = "entry", containerFactory = "replyDtoChangeListener")
+    @KafkaListener(
+        topics = [KafkaTopics.CREATE_REPLY],
+        groupId = "entry",
+        containerFactory = "createReplyDtoChangeListener"
+    )
     fun createReply(createReplyEventRequest: CreateReplyEventRequest) {
         markAsRepliedService.execute(createReplyEventRequest.questionId)
     }
