+++ src/main/kotlin/hs/kr/equus/feed/infrastructure/kafka/producer/reply/ReplyCreatedEventProducerImpl.kt
@@ -1,7 +1,7 @@
 package hs.kr.equus.feed.infrastructure.kafka.producer.reply
 
 import hs.kr.equus.feed.infrastructure.kafka.configuration.KafkaTopics
-import hs.kr.equus.user.infrastructure.kafka.dto.CreateReplyEventRequest
+import hs.kr.equus.feed.infrastructure.kafka.dto.CreateReplyEventRequest
 import org.springframework.kafka.core.KafkaTemplate
 import org.springframework.stereotype.Service
 import java.util.*
@@ -10,7 +10,7 @@ import java.util.*
 class ReplyCreatedEventProducerImpl(
     private val kafkaTemplate: KafkaTemplate<String, CreateReplyEventRequest>
 ) : ReplyCreatedEventProducer {
-    override fun send(questionId: UUID) {
-        kafkaTemplate.send(KafkaTopics.CREATE_REPLY, CreateReplyEventRequest(questionId))
+    override fun send(createEventRequest: CreateReplyEventRequest) {
+        kafkaTemplate.send(KafkaTopics.CREATE_REPLY, createEventRequest)
     }
 }
